api/pyproject.toml
[project]
name = "open-swim"
version = "0.1.0"
description = "Open Swim Python project"
readme = "README.md"
authors = [ { name = "Your Name", email = "you@example.com" } ]
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "paho-mqtt>=2.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "requests>=2.31.0"
]

[project.scripts]
open-swim = "open_swim.app:run"

[tool.uv]
# uv-specific configuration (optional tweaks)
package = true

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true



[project.optional-dependencies]
dev = [
    "mypy>=1.18.2",
    "types-requests>=2.31.0.20241016",
]
